Oops missed unit test. Remove problematic SVDRecommender constructor that does no training?
git-svn-id: https://svn.apache.org/repos/asf/lucene/mahout/trunk@834083 13f79535-47bb-0310-9956-ffa450edef68

+++ core/src/main/java/org/apache/mahout/cf/taste/impl/recommender/slopeone/MemoryDiffStorage.java
@@ -271,7 +271,7 @@ public final class MemoryDiffStorage implements DiffStorage {
     PreferenceArray userPreferences = dataModel.getPreferencesFromUser(userID);
     int length = userPreferences.length();
     for (int i = 0; i < length - 1; i++) {
-      double prefAValue = userPreferences.getValue(i);
+      float prefAValue = userPreferences.getValue(i);
       long itemIDA = userPreferences.getItemID(i);
       FastByIDMap<RunningAverage> aMap = averageDiffs.get(itemIDA);
       if (aMap == null) {
